+++ spec/lib/capistrano/hg_spec.rb
@@ -77,5 +77,14 @@ module Capistrano
         subject.release
       end
     end
+
+    describe "#fetch_revision" do
+      it "should strip trailing whitespace" do
+        context.expects(:fetch).with(:branch).returns(:branch)
+        context.expects(:capture).with(:hg, "log --rev branch --template \"{node}\n\"").returns("01abcde\n")
+        revision = subject.fetch_revision
+        expect(revision).to eq("01abcde")
+      end
+    end
   end
 end